Fixed NpcInfo and SummonInfo isFlying.

+++ L2J_Mobius_10.3_MasterClass/java/org/l2jmobius/gameserver/network/serverpackets/NpcInfo.java
@@ -355,7 +355,7 @@ public class NpcInfo extends AbstractMaskPacket<NpcInfoType>
 		}
 		if (containsMask(NpcInfoType.FLYING))
 		{
-			writeByte(_npc.isFlying());
+			writeInt(_npc.isFlying());
 		}
 		if (containsMask(NpcInfoType.CLONE))
 		{
