+++ L2J_Mobius_Essence_6.2_Vanguard/java/org/l2jmobius/gameserver/network/serverpackets/SummonInfo.java
@@ -283,7 +283,7 @@ public class SummonInfo extends AbstractMaskPacket<NpcInfoType>
 		}
 		if (containsMask(NpcInfoType.FLYING))
 		{
-			writeByte(_summon.isFlying());
+			writeInt(_summon.isFlying());
 		}
 		if (containsMask(NpcInfoType.CLONE))
 		{
